Move default-setting for hidden fields to helper
Previously the plugin would try to figure out these defaults again when
loading the form to validate it when saving the comment, but the save
URL doesn't have the right parameters, leading to a notice, plus those
defaults are useless at that point. Just load them in the view helper
only.

Also, remove all the code dealing with the "module" hidden input as its
value actually wasn't being used.